Refactor and add new use case to download a document
