+++ api/lib/UseCases/DownloadDocument.js
@@ -1,9 +1,10 @@
 module.exports = (options) => {
-  const getDocumentGateway = options.getDocumentGateway;
-  // const id = "27810857";
-  const type = "msg";
-  return async (id) => {
-    const document = await getDocumentGateway.execute(id, type);
+  const getDocument = options.getDocument;
+  const saveDocumentToS3 = options.saveDocumentToS3;
+  const id = "27810857";
+  return async () => {
+    const document = await getDocument(id);
+    await saveDocumentToS3(id, document);
     return document;
   };
 };
